third_party/blink/web_tests/editing/shadow/contenteditable-propagation-to-distributed-node-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test checks that contenteditable is propagated to distributed nodes.

PASS getComputedStyle(shadowRoot.querySelector('span'), null).webkitUserModify is "read-only"
PASS getComputedStyle(shadowRoot.querySelector('content').getDistributedNodes()[0], null).webkitUserModify is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
